Remove the `moves` package from the chess engine
